requirements.txt
opencv-python
numpy
ultralytics
Pillow
scikit-learn
webcolors
